requirements.txt
typer[all]>=0.9.0
pydantic>=2.0.0
pyyaml>=6.0
rich>=13.0.0
google-genai
python-dotenv
